refactoring files and folders
