+++ src/lib/domain/data/base/_interfaces.ts
@@ -1,1 +1,1 @@
-export * from './IDataSource';
+export * from './interfaces';
