+++ src/lib/domain/data/base/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './IDataSource'
