+++ src/lib/domain/data/indicators/_implementations.ts
@@ -1,5 +1,3 @@
 
-export * from './internals/IndicatorProvider';
-export * from './internals/IndicatorCalculationContext';
-export * from './internals/Indicator';
-export * as IndicatorsImpl from './internals/indicators/_implemenations'
+export * from './implementations/';
+export * as IndicatorsImpl from './implementations/talib-indicators'
