+++ src/lib/domain/data/indicators/_interfaces.ts
@@ -1,4 +1,1 @@
-import { IIndicator } from './internals/IIndicator';
-import { IIndicatorCalculationContext } from './internals/IIndicatorCalculationContext';
-import { IndicatorConfig } from './internals/IndicatorConfig';
-export { IIndicator, IIndicatorCalculationContext, IndicatorConfig };
+export * from './interfaces'
