+++ src/lib/domain/data/indicators/implementations/Indicator.ts
@@ -1,4 +1,4 @@
-import { IndicatorConfig } from './IndicatorConfig';
+import { IndicatorConfig } from '../interfaces/IndicatorConfig';
 
 export class Indicator {
   constructor(public cfg: IndicatorConfig) {
