+++ src/lib/domain/data/indicators/implementations/talib-indicators/MinusDi.ts
@@ -1,16 +1,16 @@
-import { IIndicator } from '../IIndicator';
-import { IIndicatorCalculationContext } from '../IIndicatorCalculationContext';
+import { IIndicator } from '../../interfaces/IIndicator';
+import { IIndicatorCalculationContext } from '../../interfaces/IIndicatorCalculationContext';
 import { Indicator } from '../Indicator';
-import { IndicatorConfig } from '../IndicatorConfig';
+import { IndicatorConfig } from '../../interfaces/IndicatorConfig';
 
 import { TalibWrapperEx } from './talibWrapper';
 
-export class PlusDi extends Indicator implements IIndicator {
+export class MinusDi extends Indicator implements IIndicator {
   constructor(cfg: IndicatorConfig, public period: number) {
     super(cfg);
   }
   async calculate(context: IIndicatorCalculationContext) {
-    const PDIArray = await TalibWrapperEx.execute({
+    const MDIArray = await TalibWrapperEx.execute({
       name: this.cfg.name,
       high: context.candleSticks.getSingleOHLCV('high'),
       low: context.candleSticks.getSingleOHLCV('low'),
@@ -19,6 +19,6 @@ export class PlusDi extends Indicator implements IIndicator {
       endIdx: context.candleSticks.items.length - 1,
       optInTimePeriod: this.period,
     });
-    context.candleSticks.addIndicator(this.cfg.id, PDIArray);
+    context.candleSticks.addIndicator(this.cfg.id, MDIArray);
   }
 }
