+++ src/lib/domain/data/indicators/implementations/talib-indicators/_interfaces.ts
@@ -1,0 +1,5 @@
+import { IndicatorConfig } from '../../interfaces/IndicatorConfig';
+import { IIndicatorCalculationContext } from '../../interfaces/IIndicatorCalculationContext';
+import { IIndicator } from '../../interfaces/IIndicator';
+
+export { IIndicator, IIndicatorCalculationContext, IndicatorConfig };
