+++ src/lib/domain/data/indicators/interfaces/IIndicator.ts
