+++ src/lib/domain/data/indicators/interfaces/IIndicatorCalculationContext.ts
