+++ src/lib/domain/data/indicators/interfaces/IndicatorConfig.ts
