--- src/lib/domain/data/indicators/internals/_interfaces.ts
@@ -1,3 +1,0 @@
-export * from './IIndicator';
-export * from './IIndicatorCalculationContext';
-export * from './IndicatorConfig';
