--- src/lib/domain/data/indicators/internals/indicators/_interfaces.ts
@@ -1,5 +1,0 @@
-import { IndicatorConfig } from '../IndicatorConfig';
-import { IIndicatorCalculationContext } from '../IIndicatorCalculationContext';
-import { IIndicator } from '../IIndicator';
-
-export { IIndicator, IIndicatorCalculationContext, IndicatorConfig };
