+++ src/lib/domain/data/stores/_interfaces.ts
@@ -1,3 +1,3 @@
-import { IDataSource } from '../base/IDataSource';
+import { IDataSource } from '../base/interfaces/IDataSource';
 export * from '../base/_interfaces';
 export type IDataStore = IDataSource
